requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
click-default-group==1.2.4
Deprecated==1.2.18
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
limits==4.0.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.45.3
tabulate==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
